fixing the project structure
