--- src/main/java/SpringBootApp/App/ProjectRepository.java
@@ -1,7 +1,0 @@
-package SpringBootApp.App;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface ProjectRepository extends CrudRepository<ProjectEntity, String> {
-
-}
