--- src/main/java/SpringBootApp/App/ProjectRequestDto.java
@@ -1,10 +1,0 @@
-package SpringBootApp.App;
-
-
-import lombok.Data;
-
-@Data
-public class ProjectRequestDto {
-    private String title;
-
-}
