--- src/main/java/SpringBootApp/App/TaskRepository.java
@@ -1,7 +1,0 @@
-package SpringBootApp.App;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface TaskRepository extends CrudRepository<TaskEntity,String> {
-
-}
