--- src/main/java/SpringBootApp/App/TaskRequestDto.java
@@ -1,14 +1,0 @@
-package SpringBootApp.App;
-
-
-import lombok.Data;
-
-@Data
-public class TaskRequestDto {
-
-    private String title;
-    private String description;
-    private Status status;
-    private String userId;
-
-}
