--- src/main/java/SpringBootApp/App/TaskUserDto.java
@@ -1,10 +1,0 @@
-package SpringBootApp.App;
-
-import lombok.Data;
-
-@Data
-public class TaskUserDto {
-
-    private String taskId;
-    private String userId;
-}
