--- src/main/java/SpringBootApp/App/UserProjectRequestDto.java
@@ -1,9 +1,0 @@
-package SpringBootApp.App;
-
-import lombok.Data;
-
-@Data
-public class UserProjectRequestDto {
-
-    private String projectId;
-}
