--- src/main/java/SpringBootApp/App/UserRepository.java
@@ -1,7 +1,0 @@
-package SpringBootApp.App;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface UserRepository extends CrudRepository<UserEntity, String> {
-
-}
