+++ src/main/java/SpringBootApp/App/payloads/UserRequestDto.java
@@ -1,8 +1,12 @@
-package SpringBootApp.App;
+package SpringBootApp.App.payloads;
 
+import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Data
+@NoArgsConstructor
+@AllArgsConstructor
 public class UserRequestDto {
 
     private String firstName;
